// src/types/patient.ts
export interface Patients {
  id: string; //cedula
  firts_name: string; //nombre
  second_name: string; //segundo nombre
  last_name: string; //apellido
  second_last_name: string; //segundo apellido
  dob: string; //fecha de nacimiento
  charge: string; //cargo
  direction: string; //direccion
  phone: string; //telefono
  email: string; //correo
  sex: string; //sexo
  age: string; //edad
  created_at: string; //fecha de consulta
}